pyproject.toml
[tool.poetry]
name = "example-project-crud"
version = "0.1.0"
description = ""
authors = ["Sereska7 <sergo_ryzakov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
alembic = "^1.13.1"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
asyncio = "^3.4.3"
uvicorn = "^0.30.1"
black = "^24.4.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
python-jose = "^3.3.0"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
